BITMAG-1063: Fixed smaller test issues, and renamed the pillar start-up script.

+++ bitrepository-reference-pillar/src/test/java/org/bitrepository/pillar/store/archive/ArchiveDirectoryTest.java
@@ -40,7 +40,10 @@ public class ArchiveDirectoryTest extends ExtendedTestCase {
     
     @AfterMethod (alwaysRun=true)
     public void shutdownTests() throws Exception {
-        FileUtils.delete(new File(DIR_NAME));
+        File dir = new File(DIR_NAME);
+        if(dir.exists()) {
+            FileUtils.delete(new File(DIR_NAME));
+        }
     }
     
     @Test( groups = {"regressiontest", "pillartest"})
